face_recognition_service/requirements.txt
Flask
face-recognition
numpy
gunicorn
python-dotenv